Split native Zenmap XML parsing helpers
