+++ zenmap/macos/native/Zenmap.swift
@@ -708,108 +708,7 @@ struct ContentView: View {
 
 
     
-    func parseNmapXML(at url: URL) -> [ScannedHost] {
-        guard let parser = XMLParser(contentsOf: url) else {
-            return []
-        }
-        
-        let delegate = NmapXMLParserDelegate()
-        parser.delegate = delegate
-        
-        guard parser.parse() else {
-            return []
-        }
-        
-        return delegate.hosts
-    }
     
-    private final class NmapXMLParserDelegate: NSObject, XMLParserDelegate {
-        private(set) var hosts: [ScannedHost] = []
-        
-        private var currentHost: ScannedHost?
-        private var currentPort: ScannedPort?
-        private var isInsideHostnames = false
-        
-        func parser(
-            _ parser: XMLParser,
-            didStartElement elementName: String,
-            namespaceURI: String?,
-            qualifiedName qName: String?,
-            attributes attributeDict: [String: String] = [:]
-        ) {
-            switch elementName {
-            case "host":
-                currentHost = ScannedHost(address: "", hostname: "", status: "unknown", ports: [])
-                
-            case "status":
-                currentHost?.status = attributeDict["state"] ?? "unknown"
-                
-            case "address":
-                if currentHost?.address.isEmpty == true {
-                    currentHost?.address = attributeDict["addr"] ?? ""
-                }
-                
-            case "hostnames":
-                isInsideHostnames = true
-                
-            case "hostname":
-                if isInsideHostnames, currentHost?.hostname.isEmpty == true {
-                    currentHost?.hostname = attributeDict["name"] ?? ""
-                }
-                
-            case "port":
-                currentPort = ScannedPort(
-                    hostAddress: currentHost?.address ?? "",
-                    protocolName: attributeDict["protocol"] ?? "",
-                    portNumber: attributeDict["portid"] ?? "",
-                    state: "unknown",
-                    serviceName: "",
-                    product: "",
-                    version: "",
-                    extraInfo: ""
-                )
-                
-            case "state":
-                currentPort?.state = attributeDict["state"] ?? "unknown"
-                
-            case "service":
-                currentPort?.serviceName = attributeDict["name"] ?? ""
-                currentPort?.product = attributeDict["product"] ?? ""
-                currentPort?.version = attributeDict["version"] ?? ""
-                currentPort?.extraInfo = attributeDict["extrainfo"] ?? ""
-                
-            default:
-                break
-            }
-        }
-        
-        func parser(
-            _ parser: XMLParser,
-            didEndElement elementName: String,
-            namespaceURI: String?,
-            qualifiedName qName: String?
-        ) {
-            switch elementName {
-            case "hostnames":
-                isInsideHostnames = false
-                
-            case "port":
-                if let currentPort {
-                    currentHost?.ports.append(currentPort)
-                }
-                currentPort = nil
-                
-            case "host":
-                if let currentHost {
-                    hosts.append(currentHost)
-                }
-                currentHost = nil
-                
-            default:
-                break
-            }
-        }
-    }
     
     
 }
